customer

+++ shop/products/forms.py
@@ -5,7 +5,7 @@ from wtforms import StringField, SubmitField, IntegerField, TextAreaField, valid
 
 class Addproducts(FlaskForm):
     name = StringField('Name',[validators.DataRequired()])
-    price = DecimalField('Price',[validators.DataRequired()])
+    price = IntegerField('Price',[validators.DataRequired()])
     discount = IntegerField('Discount', default = 0)
     stock = IntegerField('Stock',[validators.DataRequired()])
     description = TextAreaField('Description', [validators.DataRequired()])
